rearranged code
