+++ src/idpconfgen/components/bgeo_strategies/__init__.py
@@ -6,14 +6,14 @@ from idpconfgen.components.bgeo_strategies.int2cart.bgeo_int2cart import (
     )
 from idpconfgen.components.bgeo_strategies.int2cart.bgeo_int2cart import \
     name as bgeo_int2cart_name
+from idpconfgen.components.bgeo_strategies.fixed import \
+    name as bgeo_fixed_name
 from idpconfgen.components.bgeo_strategies.sampling import \
     name as bgeo_sampling_name
 
 
 bgeo_strategies_default = bgeo_sampling_name
 """The default bond geometry sampling strategy."""
-
-bgeo_fixed_name = "fixed"
 
 bgeo_strategies = (
     bgeo_fixed_name,
