+++ src/idpconfgen/core/definitions.py
@@ -179,37 +179,35 @@ blocked_ids = [
 residue_elements = {'C', 'O', 'N', 'H', 'S', 'Se', 'D'}
 minimal_bb_atoms = ['N', 'CA', 'C']  # ordered!
 
-'''
-"""
-Builder Definitions For Fixed Bond Geometries
----------------------------------------------
-Average values of the backbone angles calculated from Dunbrack PISCES
-cull_d200611/200611/cullpdb_pc90_res1.6_R0.25_d200611_chains8807
-
-Float values are represented as ratio of integers
-https://docs.python.org/3/tutorial/floatingpoint.html
-"""
-average_N_CA_C = 8731046790257777 / 4503599627370496  # +- 0.04375239960584633
-average_CA_C_Np1 = 4587708133805365 / 2251799813685248  # +- 0.022904896537130497
-average_Np1_C_O = 4733796466948169 / 2251799813685248  # +- 0.019050491268134375
-average_CA_C_O = 4825315589323725 / 2251799813685248  # +- 0.017982788310237034
-average_Cm1_N_CA = 2385749441983237 / 1125899906842624  # +- 0.029039312259214314
-bend_CA_C_OXT = 2 * pi / 3
-
-# pi corrected angles needed for the building algorithm
-build_bend_N_CA_C = (pi - average_N_CA_C) / 2
-build_bend_CA_C_Np1 = (pi - average_CA_C_Np1) / 2
-build_bend_CA_C_O = average_CA_C_O / 2  # this angle does not require `pi -`
-build_bend_Cm1_N_CA = (pi - average_Cm1_N_CA) / 2
-build_bend_CA_C_OXT = (pi - bend_CA_C_OXT) / 2
-
-distance_N_CA = 6576479998126497 / 4503599627370496  # 1.46027 +- 0.013036
-distance_CA_C = 6861872558247717 / 4503599627370496  # 1.52364 +- 0.012599
-distance_C_Np1 = 2996436734567847 / 2251799813685248  # 1.33068 +- 0.009621
-distance_C_O = 5556993099130213 / 4503599627370496  # 1.234 +- 0.0121
-distance_C_OXT = 1.27
-
-distance_N_CA_std = 0.013036529567238726
-distance_CA_C_std = 0.012599655969373144
-distance_C_Np1_std = 0.009621596711934686
-'''
+# """
+# Builder Definitions For Fixed Bond Geometries
+# ---------------------------------------------
+# Average values of the backbone angles calculated from Dunbrack PISCES
+# cull_d200611/200611/cullpdb_pc90_res1.6_R0.25_d200611_chains8807
+#
+# Float values are represented as ratio of integers
+# https://docs.python.org/3/tutorial/floatingpoint.html
+# """
+# average_N_CA_C = 8731046790257777 / 4503599627370496  # +- 0.04375239960584633
+# average_CA_C_Np1 = 4587708133805365 / 2251799813685248  # +- 0.022904896537130497
+# average_Np1_C_O = 4733796466948169 / 2251799813685248  # +- 0.019050491268134375
+# average_CA_C_O = 4825315589323725 / 2251799813685248  # +- 0.017982788310237034
+# average_Cm1_N_CA = 2385749441983237 / 1125899906842624  # +- 0.029039312259214314
+# bend_CA_C_OXT = 2 * pi / 3
+#
+# # pi corrected angles needed for the building algorithm
+# build_bend_N_CA_C = (pi - average_N_CA_C) / 2
+# build_bend_CA_C_Np1 = (pi - average_CA_C_Np1) / 2
+# build_bend_CA_C_O = average_CA_C_O / 2  # this angle does not require `pi -`
+# build_bend_Cm1_N_CA = (pi - average_Cm1_N_CA) / 2
+# build_bend_CA_C_OXT = (pi - bend_CA_C_OXT) / 2
+#
+# distance_N_CA = 6576479998126497 / 4503599627370496  # 1.46027 +- 0.013036
+# distance_CA_C = 6861872558247717 / 4503599627370496  # 1.52364 +- 0.012599
+# distance_C_Np1 = 2996436734567847 / 2251799813685248  # 1.33068 +- 0.009621
+# distance_C_O = 5556993099130213 / 4503599627370496  # 1.234 +- 0.0121
+# distance_C_OXT = 1.27
+#
+# distance_N_CA_std = 0.013036529567238726
+# distance_CA_C_std = 0.012599655969373144
+# distance_C_Np1_std = 0.009621596711934686
